@@ -15,17 +15,17 @@
   "devDependencies": {
     "@commitlint/cli": "^16.2.3",
     "@commitlint/config-conventional": "^16.2.1",
-    "@tophat/conventional-changelog-config": "^0.6.1",
+    "@tophat/conventional-changelog-config": "^0.6.2",
     "@types/eslint": "^8.4.1",
-    "@types/prettier": "^2.4.4",
-    "@typescript-eslint/eslint-plugin": "^5.17.0",
-    "@typescript-eslint/parser": "^5.17.0",
-    "eslint": "^8.12.0",
+    "@types/prettier": "^2.6.0",
+    "@typescript-eslint/eslint-plugin": "^5.21.0",
+    "@typescript-eslint/parser": "^5.21.0",
+    "eslint": "^8.14.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
     "husky": "^7.0.4",
-    "lint-staged": "^12.3.7",
-    "monodeploy": "^2.8.20",
+    "lint-staged": "^12.4.0",
+    "monodeploy": "^2.8.21",
     "prettier": "^2.6.2",
     "typescript": "^4.6.3"
   },
